# nestjs/practise-apps/04-nestjs-crud-app/backend/src/documentation/swagger.json
{"openapi":"3.0.0","paths":{"/auth/signup":{"post":{"operationId":"AuthController_signup","parameters":[],"requestBody":{"required":true,"content":{"application/x-www-form-urlencoded":{"schema":{"$ref":"#/components/schemas/AuthCredentialDto"}},"application/json":{"schema":{"$ref":"#/components/schemas/AuthCredentialDto"}},"multipart/form-data":{"schema":{"$ref":"#/components/schemas/AuthCredentialDto"}}}},"responses":{"201":{"description":"Success","content":{"application/json":{"schema":{"type":"object","properties":{"id":{"type":"integer","example":1},"email":{"type":"string","example":"user@example.com"}}}}}},"409":{"description":"Conflict - User already exists","content":{"application/json":{"schema":{"type":"object","properties":{"message":{"type":"string","example":"Conflict - User already exists"}}}}}}},"summary":"User Signup","tags":["Auth"]}},"/auth/signin":{"post":{"operationId":"AuthController_signin","parameters":[],"requestBody":{"required":true,"content":{"application/x-www-form-urlencoded":{"schema":{"$ref":"#/components/schemas/AuthCredentialDto"}},"application/json":{"schema":{"$ref":"#/components/schemas/AuthCredentialDto"}},"multipart/form-data":{"schema":{"$ref":"#/components/schemas/AuthCredentialDto"}}}},"responses":{"200":{"description":"Success","content":{"application/json":{"schema":{"type":"object","properties":{"id":{"type":"integer","example":1},"email":{"type":"string","example":"user@example.com"}}}}}},"401":{"description":"Unauthorized - Invalid credentials","content":{"application/json":{"schema":{"type":"object","properties":{"message":{"type":"string","example":"Unauthorized - Invalid credentials"}}}}}}},"summary":"User Signin","tags":["Auth"]}},"/auth/logout":{"post":{"operationId":"AuthController_logout","parameters":[],"responses":{"200":{"description":"Success","content":{"application/json":{"schema":{"type":"object","properties":{"message":{"type":"string","example":"Logged out successfully"}}}}}},"400":{"description":"Bad Request - No active session found","content":{"application/json":{"schema":{"type":"object","properties":{"message":{"type":"string","example":"Bad Request - No active session found"}}}}}}},"summary":"User Logout","tags":["Auth"]}},"/users":{"delete":{"operationId":"UsersController_DeleteUser","parameters":[],"responses":{"200":{"description":"Success","content":{"application/json":{"schema":{"type":"object","properties":{"message":{"type":"string","example":"Deleted user with ID: 123"}}}}}},"404":{"description":"User not found","content":{"application/json":{"schema":{"type":"object","properties":{"message":{"type":"string","example":"User not found"}}}}}}},"summary":"Delete current user","tags":["Users"]}},"/posts":{"post":{"operationId":"PostsController_addPost","parameters":[],"requestBody":{"required":true,"content":{"application/x-www-form-urlencoded":{"schema":{"$ref":"#/components/schemas/PostsBody"}},"application/json":{"schema":{"$ref":"#/components/schemas/PostsBody"}},"multipart/form-data":{"schema":{"$ref":"#/components/schemas/PostsBody"}}}},"responses":{"201":{"description":"Success","content":{"application/json":{"schema":{"type":"object","properties":{"id":{"type":"integer","example":1},"title":{"type":"string","example":"My first post"},"body":{"type":"string","example":"This is the content of the post"}}}}}}},"summary":"Create a new post","tags":["Posts"]},"get":{"operationId":"PostsController_fetchAllPostsByUser","parameters":[],"responses":{"200":{"description":"Success","content":{"application/json":{"schema":{"type":"object","properties":{"posts":{"type":"array","items":{"type":"object","properties":{"id":{"type":"integer","example":1},"title":{"type":"string","example":"User post title"},"body":{"type":"string","example":"User post content"}}}}}}}}}},"summary":"Fetch all posts by a user","tags":["Posts"]},"delete":{"operationId":"PostsController_deleteAllPostsByUser","parameters":[],"responses":{"200":{"description":"Success","content":{"application/json":{"schema":{"type":"object","properties":{"message":{"type":"string","example":"Deleted all posts of user: 1"}}}}}}},"summary":"Delete all posts by a user","tags":["Posts"]}},"/posts/{id}":{"patch":{"operationId":"PostsController_updatePost","parameters":[{"name":"id","required":true,"in":"path","schema":{"example":123,"type":"string"}}],"responses":{"200":{"description":"Success","content":{"application/json":{"schema":{"type":"object","properties":{"id":{"type":"integer","example":123},"title":{"type":"string","example":"Updated post title"},"body":{"type":"string","example":"Updated post content"}}}}}},"404":{"description":"Post not found","content":{"application/json":{"schema":{"type":"object","properties":{"message":{"type":"string","example":"Post not found"}}}}}}},"summary":"Update a post","tags":["Posts"]},"delete":{"operationId":"PostsController_deletePost","parameters":[{"name":"id","required":true,"in":"path","schema":{"example":123,"type":"string"}}],"responses":{"200":{"description":"Success","content":{"application/json":{"schema":{"type":"object","properties":{"message":{"type":"string","example":"Deleted post of id: 123"}}}}}},"404":{"description":"Post not found to delete","content":{"application/json":{"schema":{"type":"object","properties":{"message":{"type":"string","example":"Post not found to delete"}}}}}}},"summary":"Delete a post by ID","tags":["Posts"]},"get":{"operationId":"PostsController_fetchPost","parameters":[{"name":"id","required":true,"in":"path","schema":{"example":123,"type":"string"}}],"responses":{"200":{"description":"Success","content":{"application/json":{"schema":{"type":"object","properties":{"id":{"type":"integer","example":123},"title":{"type":"string","example":"Post title"},"body":{"type":"string","example":"Post content"}}}}}},"404":{"description":"Post not found","content":{"application/json":{"schema":{"type":"object","properties":{"message":{"type":"string","example":"Post not found"}}}}}}},"summary":"Fetch a post by ID","tags":["Posts"]}},"/posts/all":{"get":{"operationId":"PostsController_fetchAllPosts","parameters":[],"responses":{"200":{"description":"Success","content":{"application/json":{"schema":{"type":"object","properties":{"posts":{"type":"array","items":{"type":"object","properties":{"id":{"type":"integer","example":1},"title":{"type":"string","example":"Post title"},"body":{"type":"string","example":"Post content"}}}}}}}}}},"summary":"Fetch all posts","tags":["Posts"]}}},"info":{"title":"NestJS Crud App API","description":"API Documentation using RapiDoc","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"AuthCredentialDto":{"type":"object","properties":{"email":{"type":"string","example":"user@example.com"},"password":{"type":"string","example":"StrongPassword123!","minLength":6}},"required":["email","password"]},"PostsBody":{"type":"object","properties":{"title":{"type":"string","example":"My First Blog Post"},"body":{"type":"string","example":"This is the content of my blog post..."}},"required":["title","body"]}}}}